requirements.txt
streamlit==1.29.0
supabase==2.0.3
openai==1.6.1
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.17.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
PyPDF2==3.0.1
python-multipart==0.0.6
requests==2.31.0
json5==0.9.14